Rename `Annotations.h` into `CAttributes.h`
