+++ lib/HeadersGeneration/HelpersToHeader.cpp
@@ -23,8 +23,8 @@
 #include "revng/PTML/IndentedOstream.h"
 #include "revng/Pipeline/Location.h"
 #include "revng/Pipes/Ranks.h"
-#include "revng/Support/Annotations.h"
 #include "revng/Support/Assert.h"
+#include "revng/Support/CAttributes.h"
 #include "revng/Support/Debug.h"
 #include "revng/TypeNames/LLVMTypeNames.h"
