+++ lib/TypeNames/ModelTypeNames.cpp
@@ -27,8 +27,8 @@
 #include "revng/PTML/Tag.h"
 #include "revng/Pipeline/Location.h"
 #include "revng/Pipes/Ranks.h"
-#include "revng/Support/Annotations.h"
 #include "revng/Support/Assert.h"
+#include "revng/Support/CAttributes.h"
 #include "revng/TypeNames/LLVMTypeNames.h"
 #include "revng/TypeNames/ModelCBuilder.h"
